Add network client documentation
