+++ FakeNFT/Foundation/Caching/NftStorage.swift
