+++ FakeNFT/Foundation/Network/DefaultNetworkClient.swift
@@ -1,78 +1,42 @@
 import Foundation
 
-enum NetworkClientError: Error {
-    case httpStatusCode(Int)
-    case urlRequestError(Error)
-    case urlSessionError
-    case parsingError
-}
-
-protocol NetworkClient {
-    @discardableResult
-    func send(request: NetworkRequest,
-              completionQueue: DispatchQueue,
-              onResponse: @escaping (Result<Data, Error>) -> Void) -> NetworkTask?
-
-    @discardableResult
-    func send<T: Decodable>(request: NetworkRequest,
-                            type: T.Type,
-                            completionQueue: DispatchQueue,
-                            onResponse: @escaping (Result<T, Error>) -> Void) -> NetworkTask?
-}
-
-extension NetworkClient {
-
-    @discardableResult
-    func send(request: NetworkRequest,
-              onResponse: @escaping (Result<Data, Error>) -> Void) -> NetworkTask? {
-        send(request: request, completionQueue: .main, onResponse: onResponse)
-    }
-
-    @discardableResult
-    func send<T: Decodable>(request: NetworkRequest,
-                            type: T.Type,
-                            onResponse: @escaping (Result<T, Error>) -> Void) -> NetworkTask? {
-        send(request: request, type: type, completionQueue: .main, onResponse: onResponse)
-    }
-}
-
 struct DefaultNetworkClient: NetworkClient {
     private let session: URLSession
     private let decoder: JSONDecoder
     private let encoder: JSONEncoder
-
+    
     init(session: URLSession = URLSession.shared,
          decoder: JSONDecoder = JSONDecoder(),
          encoder: JSONEncoder = JSONEncoder()) {
+        
         self.session = session
         self.decoder = decoder
         self.encoder = encoder
     }
-
+    
     @discardableResult
-    func send(
-        request: NetworkRequest,
-        completionQueue: DispatchQueue,
-        onResponse: @escaping (Result<Data, Error>) -> Void
-    ) -> NetworkTask? {
+    func send(request: NetworkRequest,
+              completionQueue: DispatchQueue,
+              onResponse: @escaping (Result<Data, Error>) -> Void) -> NetworkTask? {
+        
         let onResponse: (Result<Data, Error>) -> Void = { result in
             completionQueue.async {
                 onResponse(result)
             }
         }
         guard let urlRequest = create(request: request) else { return nil }
-
+        
         let task = session.dataTask(with: urlRequest) { data, response, error in
             guard let response = response as? HTTPURLResponse else {
                 onResponse(.failure(NetworkClientError.urlSessionError))
                 return
             }
-
+            
             guard 200 ..< 300 ~= response.statusCode else {
                 onResponse(.failure(NetworkClientError.httpStatusCode(response.statusCode)))
                 return
             }
-
+            
             if let data = data {
                 onResponse(.success(data))
                 return
@@ -84,12 +48,12 @@ struct DefaultNetworkClient: NetworkClient {
                 return
             }
         }
-
+        
         task.resume()
-
+        
         return DefaultNetworkTask(dataTask: task)
     }
-
+    
     @discardableResult
     func send<T: Decodable>(
         request: NetworkRequest,
@@ -106,38 +70,38 @@ struct DefaultNetworkClient: NetworkClient {
             }
         }
     }
-
+    
     // MARK: - Private
-
+    
     private func create(request: NetworkRequest) -> URLRequest? {
         guard let endpoint = request.endpoint else {
             assertionFailure("Empty endpoint")
             return nil
         }
-
+        
         var urlRequest = URLRequest(url: endpoint)
         urlRequest.httpMethod = request.httpMethod.rawValue
-
+        
         urlRequest.addValue(RequestConstants.token, forHTTPHeaderField: "X-Practicum-Mobile-Token")
-
+        
         if let dtoDictionary = request.dto?.asDictionary() {
             var urlComponents = URLComponents()
             let queryItems = dtoDictionary.map { field in
                 URLQueryItem(
                     name: field.key,
                     value: field.value
-                    )
+                )
             }
             urlComponents.queryItems = queryItems
             urlRequest.httpBody = urlComponents.query?.data(using: .utf8)
             urlRequest.setValue("application/json", forHTTPHeaderField: "Content-Type")
         }
-
+        
         urlRequest.setValue("application/x-www-form-urlencoded", forHTTPHeaderField: "Content-Type")
-
+        
         return urlRequest
     }
-
+    
     private func parse<T: Decodable>(data: Data, type _: T.Type, onResponse: @escaping (Result<T, Error>) -> Void) {
         do {
             let response = try decoder.decode(T.self, from: data)
