+++ FakeNFT/Foundation/Network/NetworkClientError.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+enum NetworkClientError: Error {
+    case httpStatusCode(Int)
+    case urlRequestError(Error)
+    case urlSessionError
+    case parsingError
+}
