+++ FakeNFT/Foundation/Network/NetworkRequest.swift
@@ -17,7 +17,6 @@ protocol Dto {
     func asDictionary() -> [String: String]
 }
 
-// default values
 extension NetworkRequest {
     var httpMethod: HttpMethod { .get }
     var dto: Encodable? { nil }
